+++ src/main/java/net/frozenblock/lib/file/transfer/FileTransferPacket.java
@@ -17,18 +17,11 @@
 
 package net.frozenblock.lib.file.transfer;
 
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.Lists;
-import com.mojang.datafixers.util.Pair;
-import io.netty.buffer.ByteBuf;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicInteger;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.frozenblock.lib.FrozenLibConstants;
 import net.frozenblock.lib.config.frozenlib_config.FrozenLibConfig;
@@ -37,7 +30,6 @@ import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.server.level.ServerPlayer;
-import org.apache.commons.lang3.ArrayUtils;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.Unmodifiable;
@@ -49,62 +41,46 @@ import org.jetbrains.annotations.Unmodifiable;
  * @param fileName The name of the wanted file, including the file extension if transferring, excluding the file extension if requesting.
  * @param fileExtensions The names of the wanted file's possible extensions.
  * @param request Whether this is for a file request or not. If true, will cause a second transfer packet to be sent back in response with the file if possible.
- * @param snippet The raw data being transferred over the packet, as well as its index (used when split into multiple packets.) Will be ignored if this is a request.
- * @param totalPacketCount The total amount of packets to be sent for the file transfer. Will be ignored if this is a request.
+ * @param data The data to be transferred.
  */
-public record FileTransferPacket(String transferPath, String fileName, List<String> fileExtensions, boolean request, FileTransferSnippet snippet, int totalPacketCount) implements CustomPacketPayload {
+public record FileTransferPacket(String transferPath, String fileName, List<String> fileExtensions, boolean request, byte[] data) implements CustomPacketPayload {
 	public static final String LOCAL_SOURCE = ".local";
+	private static final byte[] EMPTY_DATA = new byte[0];
 	@ApiStatus.Internal
 	public static final Type<FileTransferPacket> PACKET_TYPE = new Type<>(FrozenLibConstants.id("file_transfer"));
 	@ApiStatus.Internal
 	public static final StreamCodec<FriendlyByteBuf, FileTransferPacket> STREAM_CODEC = StreamCodec.ofMember(FileTransferPacket::write, FileTransferPacket::create);
-	private static final int MAX_BYTES_PER_TRANSFER = 1835008; // 1.75MB
+	public static final int MAX_SIZE_PER_TRANSFER = 1835008; // 1.75MB
 
 	@ApiStatus.Internal
 	public static FileTransferPacket create(FriendlyByteBuf buf) {
-		return new FileTransferPacket(buf.readUtf(), buf.readUtf(), buf.readList(ByteBufCodecs.STRING_UTF8), buf.readBoolean(), FileTransferSnippet.read(buf), buf.readVarInt());
+		return new FileTransferPacket(buf.readUtf(), buf.readUtf(), buf.readList(ByteBufCodecs.STRING_UTF8), buf.readBoolean(), buf.readByteArray());
 	}
 
 	/**
-	 * Creates a {@link List} of file transfer packets.
+	 * Creates a file transfer packet.
 	 *
 	 * @param destPath The path inside Minecraft's directory to send the file to.
 	 * @param file The file to be sent.
 	 * @return A {@link List} of new file transfer packets.
-	 * @throws IOException
+	 * @throws IOException if file reading fails.
 	 */
 	@Unmodifiable
-	public static List<FileTransferPacket> create(String destPath, File file) throws IOException {
-		final Pair<Integer, List<FileTransferSnippet>> snippets = createSnippets(readFile(file));
-		final int totalPacketCount = snippets.getFirst();
-
-		final ArrayList<FileTransferPacket> packets = new ArrayList<>();
-		for (FileTransferSnippet snippet : snippets.getSecond()) {
-			packets.add(
-				new FileTransferPacket(
-					destPath,
-					file.getName(),
-					List.of(),
-					false,
-					snippet,
-					totalPacketCount
-				)
-			);
-		}
-
-		return ImmutableList.copyOf(packets);
+	public static FileTransferPacket create(String destPath, File file) throws IOException {
+		final byte[] data = readFile(file);
+		return new FileTransferPacket(destPath, file.getName(), List.of(), false, data);
 	}
 
 	/**
 	 * Create a file request packet.
 	 *
-	 * @param requestPath    The path inside Minecraft's directory the requested file should be located.
-	 * @param fileName       The requested file's name, excluding the file extension.
-	 * @param fileExtensions The names of the wanted file's possible extensions.
+	 * @param requestPath The path inside Minecraft's directory the requested file should be located.
+	 * @param fileName The requested file's name, excluding the file extension.
+	 * @param fileExtensions The possible file extensions of the requested file.
 	 * @return The new file request packet.
 	 */
 	public static FileTransferPacket createRequest(String requestPath, String fileName, List<String> fileExtensions) {
-		return new FileTransferPacket(requestPath, fileName, fileExtensions, true, FileTransferSnippet.EMPTY, 0);
+		return new FileTransferPacket(requestPath, fileName, fileExtensions, true, EMPTY_DATA);
 	}
 
 	@ApiStatus.Internal
@@ -132,7 +108,7 @@ public record FileTransferPacket(String transferPath, String fileName, List<Stri
 	 */
 	public static void sendToPlayer(File file, String destPath, ServerPlayer player) throws IOException {
 		if (!FrozenLibConfig.FILE_TRANSFER_SERVER.get()) return;
-		for (FileTransferPacket packet : create(destPath, file)) ServerPlayNetworking.send(player, packet);
+		ServerPlayNetworking.send(player, create(destPath, file));
 	}
 
 	@ApiStatus.Internal
@@ -141,44 +117,13 @@ public record FileTransferPacket(String transferPath, String fileName, List<Stri
 		buf.writeUtf(this.fileName);
 		ByteBufCodecs.STRING_UTF8.apply(ByteBufCodecs.list()).encode(buf, this.fileExtensions);
 		buf.writeBoolean(this.request);
-		this.snippet.write(buf);
-		buf.writeVarInt(this.totalPacketCount);
+		buf.writeByteArray(this.data);
 	}
 
 	@ApiStatus.Internal
 	@Override
 	public Type<? extends CustomPacketPayload> type() {
 		return PACKET_TYPE;
-	}
-
-	private static Pair<Integer, List<FileTransferSnippet>> createSnippets(byte[] bytes) {
-		final AtomicInteger index = new AtomicInteger(0);
-		final List<FileTransferSnippet> snippets = new ArrayList<>();
-
-		Lists.partition(Arrays.asList(ArrayUtils.toObject(bytes)), MAX_BYTES_PER_TRANSFER).forEach(byteChunk -> {
-			snippets.add(
-				new FileTransferSnippet(
-					ArrayUtils.toPrimitive(byteChunk.toArray(new Byte[0])),
-					index.incrementAndGet()
-				)
-			);
-		});
-
-		return Pair.of(index.get(), snippets);
-	}
-
-	public record FileTransferSnippet(byte[] bytes, int index) {
-		public static final FileTransferSnippet EMPTY = new FileTransferSnippet(new byte[0], 0);
-
-		public static FileTransferSnippet read(FriendlyByteBuf byteBuf) {
-			return new FileTransferSnippet(byteBuf.readByteArray(), byteBuf.readVarInt());
-		}
-
-		public ByteBuf write(FriendlyByteBuf byteBuf) {
-			byteBuf.writeByteArray(this.bytes);
-			byteBuf.writeVarInt(this.index);
-			return byteBuf;
-		}
 	}
 }